refactor(video): peel SkylineYear camera into 6 primitives + compose
Decompose the buildKeyframes monolith into one module per cinematographic
shot under src/scene/primitives/: establish, approachCollapse, cruise,
panoramaOrbit, peakFocus, outro. compose.ts stitches them in phase order and
sorts by frame; buildKeyframes is now a thin wrapper that builds the camera
context and delegates to composeYearCamera.

Each primitive is a faithful keyframe emitter (boundary/velocity joins layered
in next). Golden snapshot byte-identical across all 5 fixtures; 124 tests
green; typecheck clean. Removed the now-dead cruise/orbit imports from
SkylineYear (lerpSpeedRemap, peakLiftAtX, densityAt, Z_FLOOR, CRUISE_Z_MAX,
ORBIT_RZ, FLYBY_END).
