+++ video/src/scene/primitives/index.ts
@@ -5,3 +5,10 @@ export * from "./timeline";
 export * from "./constants";
 export * from "./types";
 export * from "./context";
+export * from "./establish";
+export * from "./approachCollapse";
+export * from "./cruise";
+export * from "./panoramaOrbit";
+export * from "./peakFocus";
+export * from "./outro";
+export * from "./compose";
